==> AmesHousing/PCode/01_ReadData_AmesHousing.ipynb <==
# Read raw data: Ames Housing

Contents
 - load data
 - data manipulation
   - variables to use
   - dealing with categorical variables
   - train - val - test split
 - save data

Sources:
 - http://ww2.amstat.org/publications/jse/v19n3/decock.pdf
 - https://www.kaggle.com/c/house-prices-advanced-regression-techniques/

Copyright (C) 2018 Alan Chalk  
Please do not distribute or publish without permission.

## Start_.

### Import any packages needed

# Your code here to import any packages you need
import os

import pandas as pd #data manipulation
import numpy as np #data science package (math)
import re #to deal with regex
import pickle


# print the version of numpy and pandas.  e.g. print(np.__version__)
print("pandas version: "+pd.__version__)
print("numpy version: "+np.__version__)
print("re version: "+re.__version__)
# print("pickle version: "+pickle.__version__)

### Set directories and paths

# If convenient for you, define input and PData directories.  e.g. dirRawData = "../input/"
print(os.getcwd()) #double check cwd is correct
#define input and PData directories
dirInput = "../input/"
dirPData = "../PData/"



### Load data

 - Using the dictionary below (ames_dtypes), use the read_csv function in Pandas to read AmesHousing.txt
 - The name of the dataframe you create should be df_all
 - Inspect the top few lines to make sure they look sensible
 - Note that the column names contain . to separate words and they contain capitals, e.g. Lot.Frontage.  Change all column names so that capitals are replaced with lower case and periods (.) are replaced with underscores (_).  One way to do this is 
   - Use sub from the re package to replace . with underscore (_) in all column names
   - Use the .lower method on each string to change all letters to lower case

ames_dtypes = {'Order': np.int64,
             'PID': np.int64,
             'MS.SubClass': np.object,
             'MS.Zoning': np.object,
             'Lot.Frontage': np.float64,
             'Lot.Area': np.float64,
             'Street': np.object, 
             'Alley': np.object, 
             'Lot.Shape': np.object, 
             'Land.Contour': np.object,  
             'Lot.Config': np.object,
             'Land.Slope': np.object,
             'Neighborhood': np.object,
             'Condition.1': np.object,
             'Condition.2': np.object,
             'Bldg.Type': np.object,
             'House.Style': np.object,
             'Overall.Qual': np.float64,
             'Overall.Cond': np.float64,
             'Year.Built': np.float64,
             'Year.Remod.Add': np.float64,
             'Roof.Style': np.object,
             'Roof.Matl': np.object,
             'Exterior.1st': np.object,
             'Exterior.2nd': np.object,
             'Mas.Vnr.Type': np.object,
             'Mas.Vnr.Area': np.float64,
             'Exter.Qual': np.object,
             'Exter.Cond': np.object,
             'Foundation': np.object, 
             'Bsmt.Qual': np.object,
             'Bsmt.Cond': np.object,
             'Bsmt.Exposure': np.object,
             'BsmtFin.Type.1': np.object,
             'BsmtFin.SF.1': np.float64,
             'BsmtFin.Type.2': np.object,
             'BsmtFin.SF.2': np.float64,
             'Bsmt.Unf.SF': np.float64,
             'Total.Bsmt.SF': np.float64,
             'Heating': np.object,
             'Heating.QC': np.object,
             'Central.Air': np.object,
             'Electrical': np.object,
             'X1st.Flr.SF': np.float64,
             'X2nd.Flr.SF': np.float64,
             'Low.Qual.Fin.SF': np.float64,
             'Gr.Liv.Area': np.float64, 
             'Bsmt.Full.Bath': np.float64,
             'Bsmt.Half.Bath': np.float64,
             'Full.Bath': np.float64,
             'Half.Bath': np.float64,
             'Bedroom.AbvGr': np.float64,
             'Kitchen.AbvGr': np.float64,
             'Kitchen.Qual': np.object,
             'TotRms.AbvGrd': np.float64, 
             'Functional': np.object, 
             'Fireplaces': np.float64, 
             'Fireplace.Qu': np.object,
             'Garage.Type': np.object, 
             'Garage.Yr.Blt': np.float64,
             'Garage.Finish': np.object,
             'Garage.Cars': np.float64,
             'Garage.Area': np.float64, 
             'Garage.Qual': np.object, 
             'Garage.Cond': np.object, 
             'Paved.Drive': np.object,
             'Wood.Deck.SF': np.float64,
             'Open.Porch.SF': np.float64,
             'Enclosed.Porch': np.float64, 
             'X3Ssn.Porch': np.float64,
             'Screen.Porch': np.float64,
             'Pool.Area': np.float64,
             'Fence': np.object,
             'Misc.Feature': np.object,
             'Misc.Val': np.float64,
             'Mo.Sold': np.float64, 
             'Yr.Sold': np.float64,
             'Sale.Type': np.object,
             'Sale.Condition': np.object,
             'SalePrice': np.float64}
    


#default seperator is comma, but data is seperated with spaces so change sep, set dtype as prev. defined dict
df_all = pd.read_csv(dirInput+'AmesHousing.txt',sep=' ', dtype = ames_dtypes, na_values = 'NA')
df_all.head() #inspect at first few rows

# Note that the column names contain . to separate words and they contain capitals, e.g. Lot.Frontage. 
# Change all column names so that capitals are replaced with lower case and periods (.) are replaced with 
# underscores (_). One way to do this is:
# Use sub from the re package to replace . with underscore (_) in all column names
# Use the .lower method on each string to change all letters to lower case


#df_all.columns is an index, diff to work with so let's work with values of index
columns = df_all.columns.values
#write a function to replace . with _ and convert string chars to lowercase
# test = "This.Is.A.Test" # to test function
def convert(colname):
    #replace '.' with '_'
    colname = re.sub('\.','_',colname) #need backslash to escape '.' as it is a special char
    colname = colname.lower() #change to lower case
    return colname
# print(convert(test))
#Apply function. Naively can do using forloop
# for idx, colname in enumerate(columns):
#     columns[idx] = convert(colname)
# print(columns)

#alternatively can use built in functions such as map
# ?map 
#map is an iterator and returns values one at a time, so must force it into a list
columns=list(map(convert, columns))
df_all.columns = columns #assign df_all.columns to revised columns
df_all.head() #inspect first few rows

### Store the dataset and relevant variables

 - Create a dict which contains 'df_all' only with a key of df_all, e.g. dict_ = {'df_all': df_all}
 - Use pickle.dump to save the dict to /PData/01_df_all.pickle

If you have not used pickle before, see the examples at the bottom of this page: https://docs.python.org/3/library/pickle.html

#store = pd.HDFStore(dirPData + '01_df_all.h5')
#df_all.to_hdf(store, 'df_all')
#store.close()

dict_ = {'df_all': df_all}

#declare file name for convenience
f_name = dirPData +'01_df_all.pickle'

with open(f_name, 'wb') as f: #put in file as white binary
    # Pickle the 'data' dictionary using the highest protocol available.
    pickle.dump(dict_, f) #dump dictionary to open file
    #note that it is automatically closed when we're done

del f_name #deletes cuz he doesn't like stuff hanging around? Seems unnecessary
