==> AmesHousing/PCode/02a_PrepareData_Missings.ipynb <==
# Prepare Data - Ames Housing

Contents
 - start
  - packages
  - directories and paths
 - data manipulation
   - variables to use
   - dealing with missings and outliers
 - save data

Sources:
http://ww2.amstat.org/publications/jse/v19n3/decock.pdf

Copyright (C) 2018 Alan Chalk  
Please do not distribute or publish without permission.

## Start_.

**Import any packages needed** 

import os
import numpy as np
import pandas as pd
import pickle

from sklearn import preprocessing

from matplotlib import pyplot as plt
%matplotlib inline

# AC using an image (27-25-2019) with sklearn 0.20.3

**Set directories and paths**

# Set directories
print(os.getcwd())
dirRawData = "../input/"
dirPData   = "../PData/"
dirPOutput = "../POutput/"

**Load data**

#store = pd.HDFStore(dirPData + '01_df_all.h5')
#df_all = pd.read_hdf(store, 'df_all')
#store.close()

f_name = dirPData + '01_df_all.pickle'

with open(f_name, "rb") as f: #read binary as f
    dict_ = pickle.load(f)

df_all = dict_['df_all']
del f_name, dict_


### Data Manipulation

df_all.head()

**Variables to use**

Define the variables to use and not to use.  In this project we are not going to use order or pid to predict and so they are put in the list "vars_notToUse".  

Create the following variables:
 - vars_all: an np.ndarray of column names
 - var_dep: a list containing the dependent variable ('saleprice')
 - vars_notToUse: a list of variables not to use ('order' and 'pid')
 - vars_ind: a list of variables to use being all the variables in vars_all except vars_notToUse and var_dep i.e. the independent vars


vars_all = df_all.columns.values
var_dep = ['saleprice']

vars_notToUse = ['order','pid']
#use list comprehension (see below examples)
vars_ind = [var for var in vars_all if var not in (vars_notToUse+var_dep)] #as in second list comprehension example

###example of list comprehensions
numbers = range(10)
print([number for number in numbers]) #gives list
print([number for number in numbers if number % 2 == 0]) #gives list excluding odd nums

###another example
lst = ['apple','pear','rabbit']
[word for word in lst if 't' not in word]

Create:
 - vars_ind_numeric: A list of the numeric independent variables

# df_all[vars_ind]

vars_ind_numeric = [var for var in vars_ind if df_all[var].dtype != 'object']
# df_all[vars_ind_numeric]

**Deal with missing data**

Dealing with missing data is a major topic in machine learning.  We will take the simplest approach.  Delete it.

If a feature has mostly missing values - we will simply delete the feature (i.e. that column of the data).

If a feature is mostly populated and there are just one or two records with missing values - we will delete the records.

TODO 

Carry out the following:
 - print the number of rows and columns in the data
 - find the number of missings of each feature (df_all.isnull() gives missings and you can then use .sum(axis=0) )

print(df_all.shape)
#collapse axis = 0 i.e. sum missing values,
#store as series
# df_all.isnull()
srs_missing = df_all.isnull().sum(axis=0) 
print(srs_missing[srs_missing>0]) #show which features have missing values



TODO 

- You should have found above that ['lot_frontage', 'garage_yr_blt', 'mas_vnr_area'] have a reasonable number of missings. Drop these columns (inplace).
- Remove these 3 variables from vars_ind and vars_ind_numeric
- Then delete any remaining examples with missing features
- Check the number of rows and columns of the remaining data - is it what you expect?

###DROP VARIABLES
vars_toDrop = ['lot_frontage', 'garage_yr_blt', 'mas_vnr_area']
###FOR DEBUGGING, check indiv. data types of vars to drop
# print(df_all['lot_frontage'].dtype)
# print(df_all['garage_yr_blt'].dtype)
# print(df_all['mas_vnr_area'].dtype)
###ALTERNATIVELY, check all at once
# [df_all[var].dtype for var in vars_toDrop]
df_all.drop(labels=vars_toDrop,
            axis=1,
            inplace=True)
df_all.shape

###REMOVE DROPPED VARS FROM vars_ind AND vars_ind_numeric
print(len(vars_ind))
print(len(vars_ind_numeric))
###EASIEST WAY: redefine both lists according to updated df_all
vars_ind = [var for var in vars_ind if var in df_all]
vars_ind_numeric = [var for var in vars_ind if df_all[var].dtype!='object']
###ALTERNATIVELY: use set differences
#first turn list into set, then use differences, and turn it back into a list
# list(set(vars_ind).difference(set(vars_toDrop)))
# list(set(vars_ind_numeric).difference(set(vars_toDrop)))
print(len(vars_ind))
print(len(vars_ind_numeric))


# now drop the NA
df_all.dropna(axis=0,how='any', inplace=True)
temp = df_all.isnull().sum(axis=0) 
print('Empty series?', temp[temp>0].empty) #double check all missings are gone (expect empty series)
print(df_all.shape) #check no. of cols and rows to see if match what is expected
del temp #no need to keep variable

**Remove known outliers**

See http://ww2.amstat.org/publications/jse/v19n3/decock.pdf which states:
        
> I would recommend removing any houses with more than 4000 square feet from the data set (which eliminates these five unusual observations) before assigning it to students.

TODO 

Remove the 5 examples where gr_liv_area > 4000 and check the number of rows again.

df_all = df_all[df_all['gr_liv_area']<=4000]

print(df_all.shape)

**Care is needed with the index**

- Run the line of code below - note that the index for examples (rows) is not contiguous.  Why not?

Ans: Possibly because of having dropped examples and not reset index?


print(np.where(np.diff(df_all.index.values, 1) != 1))
df_all.index.values[1339:1345]

- What is the advantage of leaving it like this?
Ans: can see easily which examples have been dropped 
- Reset the index (inplace)

df_all.reset_index(drop=True, inplace=True)
# print(np.where(np.diff(df_all.index.values, 1) != 1))
# df_all.index.values[1339:1345]

Check that missings is no longer an issue

#NOTE I ALREADY DID THIS ABOVE, BUT LEAVE PROF's CODE IN HERE ANYWAY
srs_missing = df_all.isnull().sum(axis=0)
print(srs_missing[srs_missing > 0])
del srs_missing

### Store the dataset and relevant variables

# Note: if you have run the store commands and for some reason they failed - you may need to run: store.close() 
# Note: When running the first time you do not to use: store.remove()
#store = pd.HDFStore(dirPData + '02_df_all.h5')
#store.remove('df_all')
#df_all.to_hdf(store, 'df_all')
#store.close()

dict_ = {'df_all': df_all}

f_name = dirPData + '02_df.pickle'
with open(f_name, "wb") as f: #open file, write binary
    pickle.dump(dict_, f)
del f_name, dict_

###ALSO creating a dictionary with info about variables (which not to use, which are numeric and which is dependent)
dict_ = {'vars_ind_numeric': vars_ind_numeric,
        'vars_notToUse': vars_notToUse,
        'var_dep': var_dep}
###STORE in seperate file called 02_vars
f_name = dirPData + '02_vars.pickle'
with open(f_name, "wb") as f:
    pickle.dump(dict_, f)
del f_name, dict_
