==> AmesHousing/PCode/09a_SomeBasicPython_blank.ipynb <==
## Some basic Python

Notes
 - Students in past years were missing some of the basics
 - Here we quickly go though some of the basics
 - For those who need a refersher of basic Python, the first 5 chapters of https://docs.python.org/3/tutorial/index.html would be a good start.


import numpy as np
import pandas as pd

### Missings

Bascically, I suggest you read this: https://towardsdatascience.com/navigating-the-hell-of-nans-in-python-71b12558895b



**None**

In Python, None is an object which is a built-in constant which we use to represent the absence of a value.   
Its type is NoneType.
See: https://docs.python.org/3/library/constants.html?highlight=none#None


There is only one instance of None in your Python environment.

When you type a = None, you assign the variable "a" to point to this one instance of None.


When you type b = None, you assign the variable "b" to the **same** (one and only) instance of the None constant.


Since None is a Python object, if you include it is a series, it will want to raise the dtype of the series to "object".  However you will find that Pandas changes None to np.nan (which we will see can be stored as a float64).

pd.Series([1,2,3]).dtype

# Note that at index 3 we have NaN - not None


**np.nan**

In numpy, np.nan is a special floating-point NaN value defined according to an international standard (which existed long before numpy).

np.nan is a standard (IEEE 754) floating point representation of Not A Number.

We can use np.nan for missing numeric data.  It stands for Not A Number. 


I find np.nan confusing - and there is at least one other person that does: https://towardsdatascience.com/navigating-the-hell-of-nans-in-python-71b12558895b



# Note the following
a = np.nan
b = np.nan
a is b

**Iterables** and **Iterators**

My comments that follow are based on the links below. This topic somewhat confusing (to me) so you may just be better off reading the original posts: 
 - https://stackoverflow.com/questions/9884132/what-exactly-are-iterator-iterable-and-iteration
 - https://stackoverflow.com/questions/40650082/confusion-about-iterators-and-iterables-in-python 



**Range**

a = range(5)

In Python, an "iterable" is anything that you can iterate potentially iterate over (for example on the right-side of a for-loop): 
 - for x in iterable:

It is not so easy to iterate over something.  You need to keep track of where you are and stop when you get to the end.  Python does this by creating an iterator from the iterable.  One way it does this is by calling the iter function.  And iterator has a state - it keeps track of where it is up to.

# Run this cell a few times
a_iterator.__next__()


When you use a for loop in Python over the range(5) example, Python actually creates a range_iterator using the iter() function, just like we did above. 

**zip**

When you pass things to zip, it creates a zip obect that pairs the first item of each thing together, and then the second.

The zip object is an iterator.  

It has a state.  

You can go through it.  

Once.  

Then that is it.

a = b = c = range(5)

zipped = zip(a, b, c)

for thing in zipped:
    print(thing)

for thing in zipped:
    print(thing)

What went wrong?

zipped is not an iterable.  It is an iterator. This means that
 - it has the .\_\_next\_\_() property
 - it returns its contents one at the time
 - when you get to the end - that's it!

zipped = zip(a, b, c)

zipped.__next__()

zipped.__next__()

zipped.__next__()

**The unpack operator** (*)

\* unpacks arguments passed to a function

print(a)

zipped = zip(a, b, c)

zipped = zip(a, b, c)
list(zipped)

# can we call zip on the zipped object - do we get back to where we started?
zipped = zip(a, b, c)
list(zip(zipped))

# we need to unpack the zip objec
a1, b1, c1 = 

print(a1)

**Dictionaries** aka **key-value stores**

If you listen to some of the h2o video, you will hear them talking about distributed key-value stores.  These a basically dictionaries stored over many machines.

dct_all = {'col1': (1, 2, 3),
           'col2': ('a', 'b', 'c'),
           'col3': (0.1, 0.2, 0.3)}

dct_all['col1']

df_all = pd.DataFrame(dct_all)

df_all
