==> AmesHousing/PCode/09b_h2o_PythonBooklet_blank-edited.ipynb <==
## Introduction to h2o

Source
 - Based on the h2o python booklet 
 - https://docs.h2o.ai/h2o/latest-stable/h2o-docs/booklets/PythonBooklet.pdf
 - You are recommended to read the booklet up until the Machine Learning section carefully.  You are more than welcome to read the whole booklet!
 - You are also recommended to listen to at least the first part of Amy Wang tutorial at: https://www.youtube.com/watch?v=g7drhm_SdbQ

import pandas as pd
import h2o

# At the time of writing (May 2020), the latest h2o version is 3.30.0.3|
print(h2o.__version__)

### Start h2o

h2o.connect(ip='127.0.0.1', 
            port=54321, 
            https=False)

h2o.init(ip='127.0.0.1', 
         port=54321, 
         https=False)

#h2o.cluster().shutdown()

dct_all = {'col1': (1, 2, 3),
           'col2': ('a', 'b', 'c'),
           'col3': (0.1, 0.2, 0.3)}

dct_all['col1']

df_all = pd.DataFrame(dct_all)

df_all

**h2o JVM**

h2o runs on a Java Virtual Machine.  The Python h2o module allows us to send information and requests to the machine and to ask the machine for information.

We can send the information in a Pandas DataFrame to the h2o JVM with h2o.H2OFrame(df_all)


Send the data again

h2o.H2OFrame(df_all)

Where is the data?

h2o.ls()

type(h2o.ls())

h2o.ls().iloc[0][0]

h2o.get_frame('type the Key_Frame...hex here')

**Delete all of our frames from the h2o JVM**

df_h2o_ls = h2o.ls()
df_h2o_ls

for key in df_h2o_ls['key']:
    print(key)

for key in df_h2o_ls['key'].values:
    print(key)

for key in df_h2o_ls['key']:
    h2o.remove(key)

h2o.ls()

**Send the data to the h2o JVM and define the key**

Use h2o.H2OFrame(name of data frame, destination_frame="name in h2o")

h2o.ls()

# Now it is easier to get the frame
h2o.get_frame('df_all')

h2o.remove('df_all')

**Send the data to the h2o JVM and create a "handle" for it in Python**

As above, but now: h2o_df_all = h2o.H2OFrame(...)

h2o_df_all = h2o.H2OFrame(df_all, 'df_all')

type(h2o_df_all)

This "handle" points to the h2o JVM.

We can apply some methods to it and Python knows to ask the JVM to apply the method.

For example we can send a request to the JVM to tell us the shape of the data

h2o_df_all.shape

h2o.remove('df_all')

### Methods for h2o frames

import numpy as np

mat = np.random.randn(100,4)

type(mat)

mat[0:10, 0:4]

df_all = pd.DataFrame(mat, columns=list('ABCD'))

df_all.head()

Now send the data to the h2o JVM

As before with h2o_df_all = h2o.H2OFrame(...)

h2o_df_all =  h2o.H2OFrame(df_all, 'df_all')

**head and tail methods**

Similar to Pandas - use h2o_df_all.head()

# default in Pandas is 5 obs
# df_all.head()  #pandas
h2o_df_all.head()  #h2o

# and tail is the same....
h2o_df_all.tail()

**columns attribute**

Similar to Pandas

df_all.columns

# and .columns with h2o
h2o_df_all.columns

And you can also use .types (similar to pandas .dtypes)

# or .types
h2o_df_all.types

**describe method**

Similar - kind of....

df_all.describe()

# or describe
h2o_df_all.describe()

**select columns**

By location (integer)


# try using 0 (zero) as a slice
h2o_df_all[0]

or by column name

# the column name also works, try 'A'
h2o_df_all['A']

in Pandas you have to be a little bit more explicit...

# df_all[0]  #won't work

# rather you have to give it the location as an integer.... eg .iloc[:, 0]
df_all.iloc[:,0]

Selecting multiple rows or columns is fairly obvious, eg

#slice with ['A', 'B']
h2o_df_all[['A', 'B']]

**Summing rows and columns**

df_all.sum(axis=0)

df_all.sum(axis=1)[0:10]

df_all.apply(sum, axis=0)

With an anonymous function

df_all.apply(lambda x: sum(x), axis=0)

Now in h2o

h2o_df_all.apply(lambda thing: thing.sum(), axis=0)

h2o_df_all.apply(lambda thing: thing.mean(), axis=0)

h2o_df_all.apply(lambda thing: thing.mean(), axis=1)

**Missings**

dct_missings = {'A': [1, 2, 3, np.nan],
                'B': [1, 2, 3, None],
                'C': ['a', 'b', 'c', 'NA'],
                'D': ['this', 'is', 'string', None]}


df_missings = pd.DataFrame(dct_missings)
df_missings

df_missings.loc[3, 'C']

df_missings.loc[3, 'C'] == 'NA'

df_missings.dtypes

h2o_df_missings = h2o.H2OFrame.from_python(dct_missings,
                                           column_types=['numeric', 'numeric', 'enum', 'string'], 
                                           destination_frame="df_missings")
                                
h2o_df_missings

h2o_df_missings[3, 2] == 'NA'

Note that the None value in the string column gets stored as an empty string i.e. ''

h2o_df_missings[3, 3]

h2o_df_missings[3, 3] == ''

isna() will find missing numbers.  But 
 - 'NA' is not missing - it is a string with the letters N and A in it...
 - The empty string we created with None does not show up with isna()

h2o_df_missings.isna()

It seems that when we take oparations on columns, the default is to exclude missings....

h2o_df_missings['A'].mean()

# which is the same as ...
h2o_df_missings['A'].mean(na_rm=True)

# but...
h2o_df_missings['A'].mean(na_rm=False)

Other h2o dataframe methods:
- .hist
- .countmatches
- .sub / .gsub
- .strsplit
- .rbind and .cbind
- .merge
- .group_by

h2o dataframe can also deal with date and time data

**Categorical data**

 - Known as "enum" data type 
 - This is the same as a "factor" in R
 - It seems that if we do not explicitly stored a column as a string, it will store it as "enum"

df_enum = h2o.H2OFrame({
    'A': ['foo', 'bar', 'foo', 'bar', 'foo', 'bar', 'foo', 'bar'],
    'B': ['one', 'one', 'two', 'three', 'two', 'two', 'one', 'three']})

df_enum

df_enum.types

Just as in R, we can get the levels of our factors...

df_enum['A'].levels()

df_enum['B'].levels()

Creating interactions

We will talk later about interactions in our models, but note the following ....

df_enum.interaction(['A', 'B'], pairwise=False, max_factors=100, min_occurrence=1)

df_enum.interaction(['A', 'B'], pairwise=False, max_factors=100, min_occurrence=2)

- df_all.any_factor
- is.factor() and as.factor()
- df_all[colname].levels()
- df_all.interaction(['col1','col2'], args)

h2o.cluster().shutdown()

# df_enum
