==> AmesHousing/PCode/20_Interpretability.ipynb <==
## Interpretability

<a id='top'></a>

Contents
 - start
 - load model information
 - [Variable_Importance](#Variable_Importance)
 - [Partial Dependence Plots](#PDP)
 - [Individual Conditional Expectation (ICE)](#ICE)
 - [Interactions](#interactions)
 - [LIME](#LIME)
 - [Shapley](#Shapley)
 
 
Notes: 

Sources:
 - https://christophm.github.io/interpretable-ml-book/

Copyright (C) 2019 Alan Chalk  
Please do not distribute or publish without permission.

**Start:_**

import os
import numpy as np
import pandas as pd
import pickle

import h2o
import lightgbm as lgb
from catboost import CatBoostRegressor

import seaborn as sns

import matplotlib.pyplot as plt
%matplotlib inline

**Directories and paths**

# Set directories
print(os.getcwd())
dirRawData = "../RawData/"
dirPData = "../PData/"
dirPOutput = "../POutput/"

**Settings**

font = {'size'   : 22}
plt.rc('font', **font)

**Functions**

def fn_minus_MAE(y_true, y_pred):
    return -np.round(np.mean(np.abs(y_pred - y_true)))

**Load data**

# load df_all (use the none one-hot version)
#df_all = pd.read_hdf(dirPData + '02_df_all.h5', 'df_all')
f_name = dirPData + '02_df.pickle'

with (open(f_name, "rb")) as f:
    dict_ = pickle.load(f)

df_all = dict_['df_all']

del f_name, dict_

# load the variables information
f_name = dirPData + '02_vars.pickle'
with open(f_name, "rb") as f:
    dict_ = pickle.load(f)
    
var_dep              = dict_['var_dep']
vars_ind_numeric     = dict_['vars_ind_numeric']
vars_ind_categorical = dict_['vars_ind_categorical']
vars_ind_onehot      = dict_['vars_ind_onehot']

del dict_

idx_train  = df_all['fold'].isin(range(6))
idx_val    = df_all['fold'].isin([6, 7])
idx_design = df_all['fold'].isin(range(8))
idx_test   = df_all['fold'].isin([8, 9])

y = df_all[var_dep].values.ravel()
y_train = y[idx_train]
y_val = y[idx_val]
y_design = y[idx_design]
y_test = y[idx_test]

vars_ind = vars_ind_categorical + vars_ind_numeric

# I "hope" that sklearn label encoder is deterministic 
# since I did not save the LE generated when we trained the model
from sklearn import preprocessing

df_all_lgb = df_all.copy()

for col in vars_ind_categorical:
    le = preprocessing.LabelEncoder()
    le.fit(df_all[col].values)
    df_all_lgb[col] = le.transform(df_all_lgb[col].values)


df_X_test     = df_all[idx_test][vars_ind]
df_X_test_lgb = df_all_lgb[idx_test][vars_ind]

**Start h2o**

h2o.init()

**Load data into h2o**

h2o_df_test = h2o.H2OFrame(df_all[vars_ind + var_dep + ['fold']],
                         destination_frame = 'df_all')

**Load model information**

f_name = dirPData + 'dict_predictions.pickle'

with (open(f_name, "rb")) as f:
    dict_ = pickle.load(f)

df_predictions = dict_['df_predictions']


df_predictions.head()

dict_.keys()

**Load models**

# load h2o glm
m_1a_path = dict_['m_1a_path']
glm_bst = h2o.load_model(path = m_1a_path)

# load h2o random forest
m_2a_path = dict_['m_2a_path']
rf_bst = h2o.load_model(path = m_2a_path)

# load h2o random forest
m_3a_path = dict_['m_3a_path']
xgb_bst = h2o.load_model(path = m_3a_path)

# load lightgbm 
m_3b_path = dict_['m_3b_path']
lgb_bst = lgb.Booster(model_file=dirPData + '12b_lgb_bst.txt')

# load catboost and calculated feature importances
m_3c_path = dict_['m_3c_path']
ctb_bst = CatBoostRegressor()
ctb_bst.load_model(m_3c_path)
ctb_bst_feature_importances_ = dict_['m_3c_feature_importances_']
ctb_bst_feature_names_ = dict_['m_3c_feature_names_']

<a id='Variable_Importance'></a>

## Variable Importance

[Top](#top)

Variable importance measures try to answer the question - which features are most important in the trained model.  Variable importance is typically calculated in one of three ways:
 - based on fitted model parameters (e.g. the standardised coefficients of a fitted linear model)
 - based on the structure of tree based models (such as how often a variable is split on, or the improvment to the loss during training)
 - based on how much the performance of the model deteriorates if a feature is removed (by permuting it so that it becomes meaningless) 

**h2o glm**

> **How is variable importance calculated for GLM?  For GLM, the variable importance represents the coefficient magnitudes.**

This is inconsistent with other h2o model types.  If would be nice if there was an option to create variable importance for glms based on permutation.

glm_bst.varimp_plot()

**h2o random forest**

h2o does not seem to do permutation based variable importance, even for trees:

> Variable importance is determined by calculating the relative influence of each variable: whether that variable was selected to split on during the tree building process, and how much the squared error (over all trees) improved (decreased) as a result.

 - http://docs.h2o.ai/h2o/latest-stable/h2o-docs/variable-importance.html

rf_bst.varimp_plot()

**h2o xgboost**

xgb_bst.varimp_plot()

**lightgbm**

What is returned?

From the docs:
 - importance_type : string, optional (default="split")
 - How the importance is calculated.
 - If "split", result contains numbers of times the feature is used in a model.
 - If "gain", result contains total gains of splits which use the feature.


[AC view] Variable importance based on split seems to be very likely to be useless. 

df_var_imp = pd.DataFrame(sorted(zip(lgb_bst.feature_importance(importance_type='gain'),
                                 lgb_bst.feature_name())), 
                      columns=['var_imp','feature'])

df_var_imp['var_imp'] = df_var_imp['var_imp'] / df_var_imp['var_imp'].max()

df_var_imp.sort_values(by="var_imp", ascending=False, inplace=True)

fig = plt.figure(figsize=(17, 10))
sns.barplot(x="var_imp",
            y="feature",
            data=df_var_imp[0:10])
plt.title('lightgbm feature importance (gain)')
plt.tight_layout()
plt.show()

df_var_imp = pd.DataFrame(sorted(zip(lgb_bst.feature_importance(importance_type='split'),
                                 lgb_bst.feature_name())), 
                      columns=['var_imp','feature'])

df_var_imp['var_imp'] = df_var_imp['var_imp'] / df_var_imp['var_imp'].max()

df_var_imp.sort_values(by="var_imp", ascending=False, inplace=True)

fig = plt.figure(figsize=(17, 10))
sns.barplot(x="var_imp",
            y="feature",
            data=df_var_imp[0:10])
plt.title('lightgbm feature importance (split)')
plt.tight_layout()
plt.show()

**catboost**

The docs state:

> For each feature, PredictionValuesChange shows how much on average the prediction changes if the feature value changes. The bigger the value of the importance the bigger on average is the change to the prediction value, if this feature is changed.

[AC comment] However, I do not know how this is defined or calculated.

df_var_imp = pd.DataFrame(sorted(zip(ctb_bst_feature_importances_,
                                 ctb_bst_feature_names_)), 
                      columns=['var_imp','feature'])

df_var_imp['var_imp'] = df_var_imp['var_imp'] / df_var_imp['var_imp'].max()

df_var_imp.sort_values(by="var_imp", ascending=False, inplace=True)

fig = plt.figure(figsize=(15, 10))
sns.barplot(x="var_imp",
            y="feature",
            data=df_var_imp[0:10])
plt.title('catboost feature importance (prediction values change)')
plt.tight_layout()
plt.show()

**Permutation based feature importance**
 - http://rasbt.github.io/mlxtend/user_guide/evaluate/feature_importance_permutation/#example-1-feature-importance-for-classifiers
 
 We use test data for this since we have trained on all train and val data and we would get incorrect results if we used it.

We should just be able to import this function from mlxtend but it does not seem to work - so I have just copied the latest version from github

#from mlxtend.evaluate import feature_importance_permutation

def feature_importance_permutation(X, y, predict_method,
                                   metric, num_rounds=1, seed=None):
    """Feature importance imputation via permutation importance
    Parameters
    ----------
    X : NumPy array, shape = [n_samples, n_features]
        Dataset, where n_samples is the number of samples and
        n_features is the number of features.
    y : NumPy array, shape = [n_samples]
        Target values.
    predict_method : prediction function
        A callable function that predicts the target values
        from X.
    metric : str, callable
        The metric for evaluating the feature importance through
        permutation. By default, the strings 'accuracy' is
        recommended for classifiers and the string 'r2' is
        recommended for regressors. Optionally, a custom
        scoring function (e.g., `metric=scoring_func`) that
        accepts two arguments, y_true and y_pred, which have
        similar shape to the `y` array.
    num_rounds : int (default=1)
        Number of rounds the feature columns are permuted to
        compute the permutation importance.
    seed : int or None (default=None)
        Random seed for permuting the feature columns.
    Returns
    ---------
    mean_importance_vals, all_importance_vals : NumPy arrays.
      The first array, mean_importance_vals has shape [n_features, ] and
      contains the importance values for all features.
      The shape of the second array is [n_features, num_rounds] and contains
      the feature importance for each repetition. If num_rounds=1,
      it contains the same values as the first array, mean_importance_vals.
    Examples
    -----------
    For usage examples, please see
    http://rasbt.github.io/mlxtend/user_guide/evaluate/feature_importance_permutation/
    """

    if not isinstance(num_rounds, int):
        raise ValueError('num_rounds must be an integer.')
    if num_rounds < 1:
        raise ValueError('num_rounds must be greater than 1.')

    if not (metric in ('r2', 'accuracy') or hasattr(metric, '__call__')):
        raise ValueError('metric must be either "r2", "accuracy", '
                         'or a function with signature func(y_true, y_pred).')

    if metric == 'r2':
        def score_func(y_true, y_pred):
            sum_of_squares = np.sum(np.square(y_true - y_pred))
            res_sum_of_squares = np.sum(np.square(y_true - y_true.mean()))
            r2_score = 1. - (sum_of_squares / res_sum_of_squares)
            return r2_score

    elif metric == 'accuracy':
        def score_func(y_true, y_pred):
            return np.mean(y_true == y_pred)

    else:
        score_func = metric

    rng = np.random.RandomState(seed)

    mean_importance_vals = np.zeros(X.shape[1])
    all_importance_vals = np.zeros((X.shape[1], num_rounds))

    baseline = score_func(y, predict_method(X))

    for round_idx in range(num_rounds):
        for col_idx in range(X.shape[1]):
            # we take a copy of the column to be shuffled 
            # - so that we can put it back later!
            save_col = X[:, col_idx].copy()
            rng.shuffle(X[:, col_idx])
            new_score = score_func(y, predict_method(X))
            # put the column back!
            X[:, col_idx] = save_col
            importance = baseline - new_score
            mean_importance_vals[col_idx] += importance
            all_importance_vals[col_idx, round_idx] = importance
    mean_importance_vals /= num_rounds

    return mean_importance_vals, all_importance_vals

**lightgbm permutation var imp**

imp_vals, _ = feature_importance_permutation(
    X=df_X_test_lgb.values,
    y=y_test,
    predict_method=lgb_bst.predict, 
    metric = fn_minus_MAE, # the function normally uses r^2 where bigger is better
    num_rounds=1,
    seed=2019)


df_var_imp = pd.DataFrame(sorted(zip(imp_vals,
                                 df_X_test.columns)), 
                      columns=['var_imp','feature'])

df_var_imp['var_imp'] = df_var_imp['var_imp'] / df_var_imp['var_imp'].max()

df_var_imp.sort_values(by="var_imp", ascending=False, inplace=True)

fig = plt.figure(figsize=(17, 10))
sns.barplot(x="var_imp",
            y="feature",
            data=df_var_imp[0:10])
plt.title('lightgbm feature importance (permutation)')
plt.tight_layout()
plt.show()

**catboost permutation feature importance**

imp_vals, _ = feature_importance_permutation(
    predict_method=ctb_bst.predict, 
    X=df_X_test.values,
    y=y_test,
    metric=fn_minus_MAE,
    num_rounds=1,
    seed=2019)


df_var_imp = pd.DataFrame(sorted(zip(imp_vals,
                                 df_X_test.columns)), 
                      columns=['var_imp','feature'])

df_var_imp['var_imp'] = df_var_imp['var_imp'] / df_var_imp['var_imp'].max()

df_var_imp.sort_values(by="var_imp", ascending=False, inplace=True)

fig = plt.figure(figsize=(17, 10))
sns.barplot(x="var_imp",
            y="feature",
            data=df_var_imp[0:10])
plt.title('catboost feature importance (permutation)')
plt.tight_layout()
plt.show()

<a id='PDP'></a>

## Partial Dependence Plots

[Top](#top)

# needs the one hot data
#glm_bst.partial_plot()

**PDP h2o xgboost**

xgb_bst.partial_plot(data = h2o_df_test,
                     cols = ['overall_qual'],
                     nbins=20,
                     plot=True, 
                     plot_stddev = True, 
                     figsize=(9, 6), 
                     save_to_file=dirPOutput + '20_xgb_pdp_overall_qual')

PDPs don't seem to be implemented natively in lightgbm or catboost (which has a "help wanted" on this topic).  PDPs are not that difficult to program.

**PDP random forest**

rf_bst.partial_plot(data = h2o_df_test,
                     cols = ['overall_qual'],
                     nbins=20,
                     plot=True, 
                     plot_stddev = True,
                     figsize=(9, 6), 
                     save_to_file=dirPOutput + '20_rf_pdp_overall_qual')

<a id='ICE'></a>

## Individual Conditional Expectation

[Top](#top)

Sources:
 - https://github.com/AustinRochford/PyCEbox/blob/master/notebooks/PyCEBox%20Tutorial.ipynb

!pip install pycebox

from pycebox.ice import ice, ice_plot

del df_ice
#df_ice = ice(df_X_test_lgb.sample(100), 'gr_liv_area', lgb_bst.predict, num_grid_points=200)
#df_ice = ice(df_X_test_lgb.sample(100), 'year_built', lgb_bst.predict, num_grid_points=10)
#df_ice.head()

fig, (data_ax, ice_ax) = plt.subplots(ncols=2, sharex=True, sharey=True, figsize=(16, 6))
data_ax.scatter(df_X_test_lgb['year_built'],
                y_test, 
                c='k', alpha=0.5);
data_ax.set_xlabel('year_built');
data_ax.set_ylabel('sale price');
data_ax.set_title('actual data');


ice_plot(df_ice, 
         # seems to be a bug in this class so cannot use frac_to_plot < 1
         frac_to_plot=1,
         #c='k', 
         alpha=0.25,
         color_by='gr_liv_area',
         #cmap=PuOr,
         ax=ice_ax);

ice_ax.set_xlabel('year_built');
ice_ax.set_ylabel('sale price');
ice_ax.set_title('ICE curves');

#fig.savefig(dirPOutput + '20_lgb_ice_gr_liv_area')

<a id='interactions'></a>

## Interactions

[Top](#top)

Friedman's h-statistic. Only general implementation I could find is for sklearn (sklearn-gbmi) - it could be adapted to be more general.  catboost has something built in and maybe other specific modules do too.

In general these seem to be done in two ways
 - create a statistic which measures the difference between two one way measures of feature importance and a measure of the interaction feature importance
 - extract interactions from the structure of the trees

**catboost interactions**

from catboost import Pool
lst_ifi = ctb_bst.get_feature_importance(
                               data=Pool(df_X_test, cat_features=vars_ind_categorical),
                               type='Interaction',
                               prettified=True,
                               thread_count=-1,
                               verbose=False)

type(lst_ifi)

df_ifi = pd.DataFrame(lst_ifi)

df_ifi.head()

#ctb_bst_feature_names_
var_1 = df_ifi.loc[:, 0].values
var_2 = df_ifi.loc[:, 1].values
df_ifi['var_1'] = [ctb_bst_feature_names_[x] for x in var_1]
df_ifi['var_2'] = [ctb_bst_feature_names_[x] for x in var_2]
df_ifi.head()

df_ice = ice(df_X_test.sample(100), 'gr_liv_area', ctb_bst.predict, num_grid_points=200)

fig, (data_ax, ice_ax) = plt.subplots(ncols=2, sharex=True, sharey=True, figsize=(16, 6))

data_ax.scatter(df_X_test_lgb['gr_liv_area'],
                y_test, 
                c='k', alpha=0.5);
data_ax.set_xlabel('gr_liv_area');
data_ax.set_ylabel('sale price');
data_ax.set_title('actual data');

ice_plot(df_ice, 
         # seems to be a bug in this class so cannot use frac_to_plot < 1
         frac_to_plot=1,
         #c='k', 
         alpha=0.25,
         color_by='bsmtfin_sf_1',
         #cmap=PuOr,
         ax=ice_ax);

ice_ax.set_xlabel('gr_liv_area');
ice_ax.set_ylabel('sale price');
ice_ax.set_title('ICE curves');

fig.savefig(dirPOutput + '20_cat_ice_gr_liv_area_interaction')

#!pip install eli5
#from eli5 import show_weights, show_prediction
#show_prediction(ctb_bst)

<a id='LIME'></a>

## LIME

[Top](#top)

import lime
from lime.lime_tabular import LimeTabularExplainer

vars_ind_categorical_idx = [idx for idx, e in enumerate(df_X_test.columns) if e in vars_ind_categorical]

explainer = LimeTabularExplainer(df_X_test_lgb.values, 
                                 feature_names=df_X_test.columns, 
                                 class_names=['sale_price'], 
                                 categorical_features=vars_ind_categorical_idx, 
                                 training_labels=y_test,
                                 feature_selection='lasso_path',
                                 random_state=2020,
                                 verbose=True, 
                                 mode='regression')



y_test_pred = lgb_bst.predict(df_X_test_lgb)
plt.scatter(y_test_pred, y_test)

idx_min = np.argmin(y_test_pred)
idx_qtile10 = np.argwhere(np.round(y_test_pred, -3) == np.round(np.quantile(y_test_pred, 0.1),-3))[0][0]
idx_median = np.argwhere(y_test == np.median(y_test))[0][0]
idx_max = np.argmax(y_test_pred)

i = idx_qtile10
exp = explainer.explain_instance(df_X_test_lgb.values[i],
                                 lgb_bst.predict,
                                 num_features=4)

exp.show_in_notebook(show_table=True)

print(idx_median)
i = idx_median
exp = explainer.explain_instance(df_X_test_lgb.values[i],
                                 lgb_bst.predict,
                                 num_features=4)

exp.show_in_notebook(show_table=True)

i = idx_qtile90
exp = explainer.explain_instance(df_X_test_lgb.values[i],
                                 lgb_bst.predict,
                                 num_features=4)

exp.show_in_notebook(show_table=True)

<a id="Shapley"></a>

## Shapley values

[Top](#top)

!pip install --upgrade scikit-image

import shap

# load JS visualization code to notebook
shap.initjs()

# explain the model's predictions using SHAP values
# (same syntax works for LightGBM, CatBoost, and scikit-learn models)
explainer = shap.TreeExplainer(lgb_bst)
shap_values = explainer.shap_values(df_X_test_lgb.values)


# visualize the first prediction's explanation (use matplotlib=True to avoid Javascript)
shap.force_plot(explainer.expected_value, 
                shap_values[idx_qtile10,:], 
                df_X_test_lgb.iloc[idx_qtile10,:])
